Added UI
Added a health bar and number next to it that updates with your health. Added a stamina bar that updates with your stamina. It also dissappears when your stamina is maxed out. Added an attack cooldown bar that indicates how long till you can attack again. it disappears when you can attack.