Manjit branch (#1)
* Added basic movement and looking around controls

* Added sprinting

* Added jumping

* Added Crouching and Scene

Added a scene into the "Scenes" directory for testing out movement controls

* Added Headbobbing

Added headbobbing whilst moving. More headbobbing whilst sprinting. Less  headbobbing whilst crouching.  (Makes the movement feel more realistic and immersive)

* Add slope sliding functionality and update scene

Implemented the ability for the player to slide down slopes where they would've been stuck and able to constantly jump up. Updated the main scene to include slopes for testing the sliding functionality.

* Added Zoom functionality

Zoom can either be hold down or toggle. Zoom can be used for ADS later on when adding weapons and inventory

* Added toggling functionality to Crouch and bux fixes

Made an option for crouch to be untoggled so you only stay crouched if you're holding down the key. Fixed 2 bugs with the toggle where if you spammed the keys, it would take longer to toggle back. Now it's more responsive and spamming the keys works

* Added Interaction

Implemented Interactable functionality for objects:
- Added base Interactable script with layer 9 default.
- Introduced OnFocus method to highlight objects (e.g., doors) and display "Press F to open".
- Added OnInteract method for object interactions, such as opening doors.
- OnLoseFocus method de-highlights the object and hides interaction prompt.
Note: Adjust the layer in the project or script as needed.
If you need any help with this, just message me - Manjit

* Bug Fixes For Crouching

Fixed some bugs where the player got trapped under ground when trying to uncrouch without enough space. Also seperated the crouch coroutine for toggle and nontoggle.

* Added Footsteps and Updated Scene and assets

Added footsteps for different materials based on the material they're currently on. (Add a tag to the object). Added A sky to the scene to make it look nice, Added textures and materials to use for the floors. Added sound effects for footsteps.

* Added Jumping and Landing Sounds

Works, but somewhat delayed. Not sure if it's only on my end.

* Added a health system and basic UI

Added a health system. Way to take damage, regenerate health. There is an OnDamage method that should be called with a number passed through indicating the damage. Can be used on enemy attacks if within range and looking at player

* Added Stamina

* Added Attacking.

Added basic attacking by left clicking within range of an enemy for them to take damage. Made a very basic enemy script that has 100 health and takes damage.  Enemy must have "Attackable" Layer and the layer must be selected in the first person controller

* Added UI

Added a health bar and number next to it that updates with your health. Added a stamina bar that updates with your stamina. It also dissappears when your stamina is maxed out. Added an attack cooldown bar that indicates how long till you can attack again. it disappears when you can attack.

* Changed Folder names to avoid conflicts

* build: Ignore user-specific settings

---------

Co-authored-by: Tom K <tomknighton@icloud.com>