Added button prompts
Made it so that it now displays a button prompt depending on what your interact key is and it changes based on your input device. There is a different one for keyboard, playstation and xbox. Made some scripts that i'll use later for the custom keybinds so that i can use button prompts for the display of those.
(Also changed SFX to Sound Effects)