Interact bug fixes
Made it so that if there isn't a button prompt for a key you assign to interact, it just comes up with default text. Made it so that you can no longer interact and outline objects through walls. Made it so that when you pick up a key or screwdriver etc, it remvoes the "Press e to interact prompt"