Initial inventory
Created sprites for all objects that I want to be able to be picked up so far in the game. This is done through a separate screen and running a script to take an in game screenshot of that prefab. Created a proper inventory that allows the player to store 18 items, and items are stackable. Made a UI that contains 18 inventory slots and also numbers for each item saying how many in that inventory slot. Made changes to inventory and existing UI  to make sure the UI scales well with different screen sizes. To do next: make inventory toggle on and off, let the user drag items around and move them in the inventory, implement crafting etc