Inventory + Hotbar mainly finished
Added a hotbar that I will soon allow the player to equip items from. updated it so that when you open the inventory, you can't look around or interact or attack. Made it so that you can drag items from inventory spots to other and even swap them around. Made it so that when dragging an item it uses its sprite so you can keep track of the items that move. Made it so that the slots and panel scale even better for the inventory and hotbar no matter the display size