Inventory improvements
Made improvements so now if you have an item in your hotbar, and then you pick up an item it goes to the hotbar stack if it stacks rather than creating a separate stack for it. Also made quality of life changes to let you drop items onto a stack to combine them together and any excess goes back to the original spot. Also fixed a warning with the screenshot script. Also created pickup scripts for the toothbrush, razorblade and razorbrush (melee weapon)