Inventory (#23)
* Doors + Animations

Made a door that opens both ways with animations for opening and closing

* Autoclosing doors

Doors now have a coroutine that checks ever 3 seconds after opening a door if you're out of range and then they'll automatically close

* Added Doors

Added doors all around the house at all entrances. Made it so that you can't spam interact with the door during its animations

* Changed Interactions

Made it so that interactions are more accurate since before, sometimes when you would unfocus on an object it would still stay yellow. Made it so that interactable objects now show an outline when you look at them and added a method to the interactable file that figures out the interact key (so if someone changes their keybinds it still matches)

* Added interact prompt

Interact now has "Press [E] to interact" pop that shows up when you look interactable objects. I am planning to add button prompts based on what input method is used and what the exact keybind is e.g a picture of a playstation square button or a keyboard e or f button depending on the binding

* Added button prompts

Made it so that it now displays a button prompt depending on what your interact key is and it changes based on your input device. There is a different one for keyboard, playstation and xbox. Made some scripts that i'll use later for the custom keybinds so that i can use button prompts for the display of those.
(Also changed SFX to Sound Effects)

* Button prompt alignment

Basic button prompt alignment that didn't go through on the last branch

* Interact bug fixes

Made it so that if there isn't a button prompt for a key you assign to interact, it just comes up with default text. Made it so that you can no longer interact and outline objects through walls. Made it so that when you pick up a key or screwdriver etc, it remvoes the "Press e to interact prompt"

* Models + Interact changes

Created custom models from scratch in blender, combined some of them, placed them around the house and set them up as currently "pick upable" I'm planning on making the toothbrush and the razorblade inventory items then craft a weapon with them. Fixed an interact bug where it shows interact prompt when you initially start the game. Fixed this by moving it over from UIManager to PlayerUI

* Initial inventory

Created sprites for all objects that I want to be able to be picked up so far in the game. This is done through a separate screen and running a script to take an in game screenshot of that prefab. Created a proper inventory that allows the player to store 18 items, and items are stackable. Made a UI that contains 18 inventory slots and also numbers for each item saying how many in that inventory slot. Made changes to inventory and existing UI  to make sure the UI scales well with different screen sizes. To do next: make inventory toggle on and off, let the user drag items around and move them in the inventory, implement crafting etc

* Updated scene

Mainly doing this commit to test some stuff with the branch

* Inventory toggle

Added a player input action that allows the user to toggle the inventory on and off using tab

* Inventory + Hotbar mainly finished

Added a hotbar that I will soon allow the player to equip items from. updated it so that when you open the inventory, you can't look around or interact or attack. Made it so that you can drag items from inventory spots to other and even swap them around. Made it so that when dragging an item it uses its sprite so you can keep track of the items that move. Made it so that the slots and panel scale even better for the inventory and hotbar no matter the display size

* Inventory improvements

Made improvements so now if you have an item in your hotbar, and then you pick up an item it goes to the hotbar stack if it stacks rather than creating a separate stack for it. Also made quality of life changes to let you drop items onto a stack to combine them together and any excess goes back to the original spot. Also fixed a warning with the screenshot script. Also created pickup scripts for the toothbrush, razorblade and razorbrush (melee weapon)