Equippable items
Made it so that you can now "Equip" items. They don't do anything as of yet, other than just appearing in the bottom right of the players screen. I will look to make it so that if you want to interact with a door and need to use a key, you need to move it to the hotbar and then have it currently equipped to interact