Settings page (#30)
* Interact changes

Made it so that now if you want to interact with a door using an object e.g. the screwdriver, you have to equip it and then use it on the door

* Settings page

Settings page is mainly finished. Volume and sensitivity works. Still need to work on rebinding, having some issues with making sure it is persistent

* Rebinding finally works

Made a separate canvas for the settings page so that I can add a script with a dont destroy on load to it so that it persists onto the next scene where I will soon create a pause menu where the user can access the settings menu.

* Increased scroll speed

Increased scroll speed for the scroll view

* Button prompts update

Button prompts now work for the edited key bind. Removed all unnecessary debug statements too

* OcclusionCulling

Did some occlusion culling to increase performance by a lot.

* Button prompts work with rebinds

FINALLY managed to get the button prompts to work with the rebinds after working on it for 12 hours straight :sob: Made it so that it compares the playerRef to the default input system, that way you can quite easily change keybinds and have them pop up whilst also maintaining button prompts for controller too.

* Pause Menu

Completely implemented a pause menu. Escape to pause the game. you can resume it. Settings page based off the one in the main menu where you can edit volume, sensitivity and keybinds. Button prompt also updated to work if you change keybinds in game. Also fixed a few UI issues i found