Final features (#33)
* Auto Crafting

Implemented automatic crafting. It now checks the inventory if the relevant items are inside of the inventory or hotbar and if they are, it removes them from the inventory and replaces it with the craftable item.

* All Features Completed

Finished all final bugs and features. Changed all UI to be fluid and consistent. Fixed OcclusionCulling. Added quit button for desktop builds to close the game

* Comments

Comments